Revert "Use preferred names for tlf names - user's own username first"
