+++ libdokan/dir.go
@@ -21,9 +21,8 @@ type Folder struct {
 	fs   *FS
 	list *FolderList
 
-	handleMu       sync.RWMutex
-	h              *libkbfs.TlfHandle
-	hPreferredName libkbfs.PreferredTlfName
+	handleMu sync.RWMutex
+	h        *libkbfs.TlfHandle
 
 	folderBranchMu sync.Mutex
 	folderBranch   libkbfs.FolderBranch
@@ -52,14 +51,12 @@ type Folder struct {
 	noForget bool
 }
 
-func newFolder(fl *FolderList, h *libkbfs.TlfHandle,
-	hPreferredName libkbfs.PreferredTlfName) *Folder {
+func newFolder(fl *FolderList, h *libkbfs.TlfHandle) *Folder {
 	f := &Folder{
-		fs:             fl.fs,
-		list:           fl,
-		h:              h,
-		hPreferredName: hPreferredName,
-		nodes:          map[libkbfs.NodeID]dokan.File{},
+		fs:    fl.fs,
+		list:  fl,
+		h:     h,
+		nodes: map[libkbfs.NodeID]dokan.File{},
 	}
 	return f
 }
@@ -67,7 +64,7 @@ func newFolder(fl *FolderList, h *libkbfs.TlfHandle,
 func (f *Folder) name() libkbfs.CanonicalTlfName {
 	f.handleMu.RLock()
 	defer f.handleMu.RUnlock()
-	return libkbfs.CanonicalTlfName(f.hPreferredName)
+	return f.h.GetCanonicalName()
 }
 
 func (f *Folder) setFolderBranch(folderBranch libkbfs.FolderBranch) error {
@@ -152,45 +149,22 @@ func (f *Folder) BatchChanges(ctx context.Context, changes []libkbfs.NodeChange)
 }
 
 // TlfHandleChange is called when the name of a folder changes.
-// Note that newHandle may be nil. Then the handle in the folder is used.
-// This is used on e.g. logout/login.
 func (f *Folder) TlfHandleChange(ctx context.Context,
 	newHandle *libkbfs.TlfHandle) {
-	f.fs.log.CDebugf(ctx, "TlfHandleChange called on %q",
-		canonicalNameIfNotNil(newHandle))
-
 	// Handle in the background because we shouldn't lock during
 	// the notification
 	f.fs.queueNotification(func() {
-		cuser, err := libkbfs.GetCurrentUsernameIfPossible(ctx, f.fs.config.KBPKI(), f.list.public)
-		// Here we get an error, but there is little that can be done.
-		// cuser will be empty in the error case in which case we will default to the
-		// canonical format.
-		if err != nil {
-			f.fs.log.CDebugf(ctx, "tlfHandleChange: GetCurrentUserIfPossible failed: %v", err)
-		}
-		oldName, newName := func() (libkbfs.PreferredTlfName, libkbfs.PreferredTlfName) {
+		oldName := func() libkbfs.CanonicalTlfName {
 			f.handleMu.Lock()
 			defer f.handleMu.Unlock()
-			oldName := f.hPreferredName
-			if newHandle != nil {
-				f.h = newHandle
-			}
-			f.hPreferredName = f.h.GetPreferredFormat(cuser)
-			return oldName, f.hPreferredName
+			oldName := f.h.GetCanonicalName()
+			f.h = newHandle
+			return oldName
 		}()
 
-		if oldName != newName {
-			f.list.updateTlfName(ctx, string(oldName), string(newName))
-		}
+		f.list.updateTlfName(ctx, string(oldName),
+			string(newHandle.GetCanonicalName()))
 	})
-}
-
-func canonicalNameIfNotNil(h *libkbfs.TlfHandle) string {
-	if h == nil {
-		return "(nil)"
-	}
-	return string(h.GetCanonicalName())
 }
 
 func (f *Folder) resolve(ctx context.Context) (*libkbfs.TlfHandle, error) {
