+++ libdokan/folderlist.go
@@ -8,7 +8,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/kbfs/dokan"
 	"github.com/keybase/kbfs/libkbfs"
 	"golang.org/x/net/context"
@@ -126,7 +125,7 @@ func (fl *FolderList) open(ctx context.Context, oc *openContext, path []string) 
 			continue
 		}
 
-		h, err := libkbfs.ParseTlfHandlePreferred(
+		h, err := libkbfs.ParseTlfHandle(
 			ctx, fl.fs.config.KBPKI(), name, fl.public)
 		fl.fs.log.CDebugf(ctx, "FL Lookup continuing -> %v,%v", h, err)
 		switch err := err.(type) {
@@ -162,11 +161,7 @@ func (fl *FolderList) open(ctx context.Context, oc *openContext, path []string) 
 		}
 
 		fl.fs.log.CDebugf(ctx, "FL Lookup adding new child")
-		cuser, err := libkbfs.GetCurrentUsernameIfPossible(ctx, fl.fs.config.KBPKI(), h.IsPublic())
-		if err != nil {
-			return nil, false, err
-		}
-		child = newTLF(fl, h, h.GetPreferredFormat(cuser))
+		child = newTLF(fl, h)
 		fl.lockedAddChild(name, child)
 		return child.open(ctx, oc, path[1:])
 	}
@@ -184,7 +179,7 @@ func (fl *FolderList) FindFiles(ctx context.Context, fi *dokan.FileInfo, ignored
 	fl.fs.logEnter(ctx, "FL FindFiles")
 	defer func() { fl.fs.reportErr(ctx, libkbfs.ReadMode, err) }()
 
-	cuser, _, err := fl.fs.config.KBPKI().GetCurrentUserInfo(ctx)
+	_, _, err = fl.fs.config.KBPKI().GetCurrentUserInfo(ctx)
 	isLoggedIn := err == nil
 
 	var favs []libkbfs.Favorite
@@ -201,14 +196,8 @@ func (fl *FolderList) FindFiles(ctx context.Context, fi *dokan.FileInfo, ignored
 		if fav.Public != fl.public {
 			continue
 		}
-		pname, err := libkbfs.FavoriteNameToPreferredTLFNameFormatAs(cuser,
-			libkbfs.CanonicalTlfName(fav.Name))
-		if err != nil {
-			fl.fs.log.Errorf("FavoriteNameToPreferredTLFNameFormatAs: %q %v", fav.Name, err)
-			continue
-		}
 		empty = false
-		ns.Name = string(pname)
+		ns.Name = fav.Name
 		err = callback(&ns)
 		if err != nil {
 			return err
@@ -262,22 +251,5 @@ func clearFolderListCacheLoop(ctx context.Context, r *Root) {
 		}
 		r.private.clearAliasCache()
 		r.public.clearAliasCache()
-	}
-}
-
-// update things after user changed.
-func (fl *FolderList) userChanged(ctx context.Context, _, _ libkb.NormalizedUsername) {
-	var fs []*Folder
-	func() {
-		fl.mu.Lock()
-		defer fl.mu.Unlock()
-		for _, tlf := range fl.folders {
-			if tlf, ok := tlf.(*TLF); ok {
-				fs = append(fs, tlf.folder)
-			}
-		}
-	}()
-	for _, f := range fs {
-		f.TlfHandleChange(ctx, nil)
 	}
 }
