+++ libdokan/fs.go
@@ -12,7 +12,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/logger"
 	"github.com/keybase/kbfs/dokan"
 	"github.com/keybase/kbfs/dokan/winacl"
@@ -70,7 +69,7 @@ func NewFS(ctx context.Context, config libkbfs.Config, log logger.Logger) (*FS, 
 
 	ctx = wrapContext(ctx, f)
 
-	f.remoteStatus.Init(ctx, f.log, f.config, f)
+	f.remoteStatus.Init(ctx, f.log, f.config)
 	f.notifications.LaunchProcessor(ctx)
 	go clearFolderListCacheLoop(ctx, f.root)
 
@@ -497,7 +496,7 @@ func (f *FS) folderListRename(ctx context.Context, fl *FolderList, oc *openConte
 	}
 	dstName := dstPath[len(dstPath)-1]
 	// Yes, this is slow, but that is ok here.
-	if _, err := libkbfs.ParseTlfHandlePreferred(
+	if _, err := libkbfs.ParseTlfHandle(
 		ctx, f.config.KBPKI(), dstName, fl.public); err != nil {
 		return dokan.ErrObjectNameNotFound
 	}
@@ -565,15 +564,6 @@ func (f *FS) stringReadFile(contents string) dokan.File {
 		fs: f,
 	}
 }
-
-// UserChanged is called from libfs.
-func (f *FS) UserChanged(ctx context.Context, oldName, newName libkb.NormalizedUsername) {
-	f.log.CDebugf(ctx, "User changed: %q -> %q", oldName, newName)
-	f.root.public.userChanged(ctx, oldName, newName)
-	f.root.private.userChanged(ctx, oldName, newName)
-}
-
-var _ libfs.RemoteStatusUpdater = (*FS)(nil)
 
 // Root represents the root of the KBFS file system.
 type Root struct {
