+++ libdokan/tlf.go
@@ -27,9 +27,8 @@ type TLF struct {
 	emptyFile
 }
 
-func newTLF(fl *FolderList, h *libkbfs.TlfHandle,
-	name libkbfs.PreferredTlfName) *TLF {
-	folder := newFolder(fl, h, name)
+func newTLF(fl *FolderList, h *libkbfs.TlfHandle) *TLF {
+	folder := newFolder(fl, h)
 	tlf := &TLF{
 		folder: folder,
 	}
@@ -213,12 +212,13 @@ func (tlf *TLF) CanDeleteDirectory(ctx context.Context, fi *dokan.FileInfo) (err
 func (tlf *TLF) Cleanup(ctx context.Context, fi *dokan.FileInfo) {
 	var err error
 	if fi != nil && fi.IsDeleteOnClose() {
-		tlf.folder.handleMu.Lock()
-		fav := tlf.folder.h.ToFavorite()
-		tlf.folder.handleMu.Unlock()
-		tlf.folder.fs.log.CDebugf(ctx, "TLF Removing favorite %q", fav.Name)
+		name := string(tlf.folder.name())
+		tlf.folder.fs.log.CDebugf(ctx, "TLF Removing favorite %q", name)
 		defer tlf.folder.reportErr(ctx, libkbfs.WriteMode, err)
-		err = tlf.folder.fs.config.KBFSOps().DeleteFavorite(ctx, fav)
+		err = tlf.folder.fs.config.KBFSOps().DeleteFavorite(ctx, libkbfs.Favorite{
+			Name:   name,
+			Public: tlf.isPublic(),
+		})
 	}
 
 	if tlf.refcount.Decrease() {
