+++ libkbfs/interfaces.go
@@ -275,9 +275,6 @@ type KBFSOps interface {
 	// PushConnectionStatusChange updates the status of a service for
 	// human readable connection status tracking.
 	PushConnectionStatusChange(service string, newStatus error)
-	// PushStatusChange causes Status listeners to be notified via closing
-	// the status channel.
-	PushStatusChange()
 }
 
 // KeybaseService is an interface for communicating with the keybase
