+++ libkbfs/keybase_service_util.go
@@ -32,7 +32,6 @@ func serviceLoggedIn(ctx context.Context, config Config, name string,
 	config.MDServer().RefreshAuthToken(ctx)
 	config.BlockServer().RefreshAuthToken(ctx)
 	config.KBFSOps().RefreshCachedFavorites(ctx)
-	config.KBFSOps().PushStatusChange()
 }
 
 // serviceLoggedIn should be called when the current user logs out.
@@ -44,5 +43,4 @@ func serviceLoggedOut(ctx context.Context, config Config) {
 	config.MDServer().RefreshAuthToken(ctx)
 	config.BlockServer().RefreshAuthToken(ctx)
 	config.KBFSOps().RefreshCachedFavorites(ctx)
-	config.KBFSOps().PushStatusChange()
 }
