+++ libkbfs/md_ops.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/logger"
-	"github.com/keybase/client/go/protocol/keybase1"
 	"golang.org/x/net/context"
 )
 
@@ -196,11 +195,13 @@ func (md *MDOpsStandard) processMetadata(ctx context.Context,
 		return ImmutableRootMetadata{}, md.convertVerifyingKeyError(ctx, rmds, handle, err)
 	}
 
-	// Get the UID unless this is a public tlf - then proceed with empty uid.
-	var uid keybase1.UID
-	if !handle.IsPublic() {
-		_, uid, err = md.config.KBPKI().GetCurrentUserInfo(ctx)
-		if err != nil {
+	_, uid, err := md.config.KBPKI().GetCurrentUserInfo(ctx)
+	if err != nil {
+		// If this is a public folder, it's ok to proceed if we have
+		// no current session.
+		if _, ok := err.(NoCurrentSessionError); ok && !handle.IsPublic() {
+			return ImmutableRootMetadata{}, err
+		} else if !ok {
 			return ImmutableRootMetadata{}, err
 		}
 	}
